+++ Controller/OrderAddressesController.php
@@ -1,4 +1,13 @@
 <?php
 class OrderAddressesController extends CartAppController {
 
+/**
+ * Index
+ *
+ * @return void
+ */
+	public function index() {
+		$this->set('addresses', $this->OrderAddress->byUserId($this->Auth->user('id')));
+	}
+
 }