+++ Model/OrderAddress.php
@@ -1,11 +1,21 @@
 <?php
 App::uses('CartAppModel', 'Cart.Model');
+
 /**
  * Order Address Model
  *
  * Thought to be used for shipping and billing addresses
  */
 class OrderAddress extends CartAppModel {
+
+/**
+ * Behaviors
+ *
+ * @var array
+ */
+	public $actsAs = array(
+		'Search.Searchable');
+
 /**
  * belongsTo associations
  *
@@ -43,5 +53,35 @@ class OrderAddress extends CartAppModel {
 			'notEmpty' => array(
 			'rule' => 'notEmpty')),
 	);
+
+/**
+ * Filters args for search
+ *
+ * @var array
+ */
+	public $filterArgs = array(
+		array('name' => 'first_name', 'type' => 'like', 'field' => 'first_name'),
+		array('name' => 'last_name', 'type' => 'like', 'field' => 'last_name'),
+		array('name' => 'street', 'type' => 'like', 'field' => 'street'),
+		array('name' => 'zip', 'type' => 'like', 'field' => 'zip'),
+		array('name' => 'city', 'type' => 'like', 'field' => 'city'),
+	);
+
+/**
+ * Gets all addresses by user id ordered by primary address first and then first
+ * name ascendencing
+ *
+ * @param string $userId
+ * @return array
+ */
+	public function byUserId($userId = null) {
+		$this->find('all', array(
+			'contain' => array(),
+			'conditions' => array(
+				$this->alias . '.user_id' => $userId ),
+			'order' => array(
+				'primary' => 'ASC',
+				'first_name' => 'ASC')));
+	}
 
 }