+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2013-08-23
+
+ * Decoupled the checkout process from the cart, moved the related methods from the CartsController into the new CheckoutController
+ * Added an event that is triggered when an order record is changed
+
 ## 2013-04-19
 
  * moved CartManager::_initializeCart() from initialize() to startup()